Add a custom capability to wait before executing a new action
Add a new capability "lockStep":boolean. When set, the remote webdriver server
will poll a server every second to decide when the next action can be send.
New session is never delayed.

When a session will be either terminated (QUIT) or closed (CLOSE), an HTTP
request is sent first to the endpoint, then the server will wait for the ready
flag to be true.

At the moment, the endpoint is hardcoded to localhost:8888.