SYNTH-1310 : Added a new flag to flush all sessions
The new flag will allow us to explicitly flush all sessions
and then create a new one. This will avoid the problem of reference
counting and possibly having an extra session lying around.

Testing notes -
- Created and used with the synth agent. Made sure measurements ran correctly
and the logs showed that the sessions were flushed and a new session
was created.