SYNTH-1714 : The selenium server wasnt flushing the sessions correctly
The server was previously just removing it from the server's internal
data structure keeping track of the active sessions. We were removing
session from here but we werent calling QUIT on the actual driver
server. Changed the code to do this now.

Testing notes -
- Ran a sample script which created 3 drivers. The first two had
flushSessions set as True. Made sure that the first browser
and chromedriver intance was deleted when the second ones were
created.