SYNTH-3329 Remove info about credentials from webdriver logs
The selenium server logs all commands sent to the wendrivers. These
include the get commands which have the url being fetched and the
send keys command which might contain credentials. This creates a
possible security issue.

This patch will add logic which redacts whatever is being sent as
part of the send_keys command as well as remove any basic auth
credentials which might appear in the url when executing a get command.

Testing notes -
- tested manually by running a URL and a scripted job and then
manually making sure that the webdriver-server.log does not contain
any information in the get log and the send key logs are REDACTED