SYNTH-3725 Captured logs from IEDriverServer
When InternetExplorerDriver sets up a new service, it looks for the
passed capabilities. If the capabilities contain instructions on
logging, only then the driver sets up logging.

Since we want to enable logging by default, we can modify the
setupService() method so that the IEDriver is always started with
logging enabled.

Testing notes:
- Tested on Windows VM that logs from IEDriverServer are captured in a
  file driver.log inside the task directory