SYNTH-4156 Fix NPE during QUIT
Occasionally `driver.quit` encounters an error but the logic that
constructs the error message in turn encounters an NPE and as a result
we are unable to know what exactly went wrong. This commit fixes the bug
in the error message construction and hopefully help us identify the
underlying issue.

Testing notes:
- Ran @sanity on Chrome with these changes.